// MSSolutionsUI/src/app/models/claim.model.ts
export interface Claim {
  id: number;
  technicianName: string | null;
  lmStatus: string | null;
  itemName: string;
  rating: number | null;
  l3ApprovalDate: string | null;
  servicePartner: number;
  otpRequired: boolean;
  displayName: string;
  reasons: string | null;
  callCloseRemarks: string | null;
  concern: string;
  claimRegisteredPhoneNumber: string | null;
  reasonOfLoss: string | null;
  estimationFirst: number | null;
  solvyPan: string;
  l2ServiceRejectionRemarks: string | null;
  servicePartnerName: string;
  voiceUrl: string | null;
  estimationParts: number | null;
  customerAddress: string;
  brandComplaintNumber: string | null;
  address: number;
  otpVerificationTimestamp: string | null;
  warrantyStartDate: string;
  channelId: number;
  phone: string | null;
  itemPrice: number;
  warrantyDuration: number;
  invoiceDate: string | null;
  assignedExecutiveName: string | null;
  alternateContact: string | null;
  planSoldDate: string;
  raiseAdditional: boolean;
  warrantyTypeId: number;
  customerState: string;
  itemSerialNumber: string | null;
  solvyAddress: string | null;
  endDate: string | null;
  itemBrand: string;
  l2ServiceApprovalDate: string | null;
  claimVerificationDocumentUploadTimestamp: string | null;
  customerEmail: string;
  channelName: string;
  claimRedeemStatus: string | null;
  createdTime: string;
  endStatus: string | null;
  l2InsuranceRejectionRemarks: string | null;
  callAbortDateTime: string | null;
  servicePartnerCity: string;
  estimationDateTime: string | null;
  estimationActionDate: string | null;
  repairCompletionDate: string | null;
  warrantyActivationCode: string;
  invoiceActionDate: string | null;
  status: number;
  itemImei: string | null;
  hsnRequired: boolean;
  pincode: number;
  estimationLast: number | null;
  servicePartnerPhone: string;
  itemCategory: string;
  technician: string | null;
  startDate: string | null;
  customerAlternateEmail: string | null;
  servicePartnerState: string | null;
  claimRedeemedDate: string | null;
  l2InsuranceApprovalDate: string | null;
  utrNumber: string | null;
  extraInfo: string | null;
  appointmentConfirmationTime: string | null;
  created: string;
  claimRedeemed: boolean;
  previousVisitProductSerialNumber: string | null;
  servicePartnerGST: string;
  callAbortReason: string | null;
  insuredCompany: string | null;
  solvyGST: string;
  paymentDone: string;
  servicePartnerPincode: number;
  claimApproved: number;
  appointmentPendingReason: string | null;
  insured: string;
  statusName: string;
  selfInvoice: string | null;
  servicePartnerPAN: string;
  servicePartnerAddress: string;
  storeName: string;
  l1ApprovalDate: string | null;
  previousStatus: number;
  firstVisitTAT: string | null;
  paymentStatus: string | null;
  estimationPaymentAmount: number | null;
  customerPhone: string;
  warranty: number;
  claimPaymentStatus: string | null;
  registeredBy: number;
  l1RejectionRemarks: string | null;
  dropDate: string | null;
  cta: string | null;
  customerCity: string;
  item: number;
  isAddressEditable: boolean;
  registeredByName: string;
  l3RejectionRemarks: string | null;
  createdDate: string;
  claimType: number;
  repairDoneTAT: string | null;
  warrantyType: string;
  pickupDate: string | null;
  itemCategoryId: number;
  paymentId: string | null;
  remarks: string;
  type: string | null;
  claimAmount: number | null;
  appointment: number;
  solvyStateCode: string;
  paymentTxnId: string | null;
  paymentAmount: number | null;
  amount: number | null;
  invoiceNumber: string | null;
  customerName: string;
  sgst: string;
  cgst: string;
  igst: string;
}